backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
executing==2.1.0
fastapi==0.115.5
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
openai==1.56.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pipenv==2024.3.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.32.1
virtualenv==20.27.1
wcwidth==0.2.13
Werkzeug==3.1.3
